requirements.txt
fastapi
uvicorn
requests
langchain
faiss-cpu
sentence-transformers
pydantic
python-dotenv
